unsigned

+++ src/wavelet/wavthresh.c
@@ -62,7 +62,7 @@ static void wavelet_thresh_apply(const operator_data_t* _data, float mu, complex
 			if (MD_IS_SET(data->flags, i)) {
 
 				int levels = wavelet_num_levels(data->N, MD_BIT(i), data->dims, data->minsize, data->flen);
-				shift[i] = rand_range_state(data->rand_state, (1 << levels) + 1); // +1, as we want to include the limit
+				shift[i] = rand_range_state(data->rand_state, (1 << levels) + 1u); // +1, as we want to include the limit
 
 				assert(shift[i] < data->dims[i]);
 			}
@@ -75,7 +75,7 @@ static void wavelet_thresh_apply(const operator_data_t* _data, float mu, complex
 
 
 
-void wavthresh_rand_state_set(const struct operator_p_s* op, int x)
+void wavthresh_rand_state_set(const struct operator_p_s* op, unsigned long long x)
 {
 	auto data = CAST_DOWN(wavelet_thresh_s, operator_p_get_data(op));
 	rand_state_update(data->rand_state, x);
